refactor(db): extract types, schema, and migrations into sibling modules
Round 1/3 — refs #129
pattern: extract-module (mechanical)

Splits the 1851-LoC ClawStashDB monolith along the axes that have zero
behavioural coupling: type declarations, the initial CREATE TABLE block,
and the MIGRATIONS array. Public API of ClawStashDB is unchanged; db.ts
re-exports all types so every existing import keeps working.

- db-types.ts: 24 interfaces / type aliases (~220 LoC)
- db-schema.ts: INITIAL_SCHEMA_SQL constant (~85 LoC)
- db-migrations.ts: MIGRATIONS array + applyPendingMigrations() (~250 LoC)

db.ts shrinks from 1851 to 1375 LoC (-26%). No route handler, MCP file,
component, or singleton was touched. All existing tests pass unchanged.
