+++ src/server/db.ts
@@ -4,420 +4,62 @@ import crypto from 'crypto';
 import path from 'path';
 import fs from 'fs';
 import { detectLanguage } from './detect-language';
+import { INITIAL_SCHEMA_SQL } from './db-schema';
+import { applyPendingMigrations } from './db-migrations';
+import type {
+  Stash,
+  StashFile,
+  StashVersion,
+  StashVersionFile,
+  StashVersionListItem,
+  StashFileInfo,
+  StashListItem,
+  StashMeta,
+  AccessLogEntry,
+  TokenScope,
+  ApiToken,
+  ApiTokenListItem,
+  CreateStashInput,
+  UpdateStashInput,
+  ListStashesOptions,
+  SearchStashItem,
+  SearchStashesResult,
+  TagGraphOptions,
+  TagGraphResult,
+  StashGraphOptions,
+  StashGraphNode,
+  StashGraphEdge,
+  StashGraphResult,
+} from './db-types';
 
-export interface StashFile {
-  id: string;
-  stash_id: string;
-  filename: string;
-  content: string;
-  language: string;
-  sort_order: number;
-}
-
-export interface Stash {
-  id: string;
-  name: string;
-  description: string;
-  tags: string[];
-  metadata: Record<string, unknown>;
-  version: number;
-  archived: boolean;
-  created_at: string;
-  updated_at: string;
-  files: StashFile[];
-}
-
-export interface StashVersionFile {
-  filename: string;
-  content: string;
-  language: string;
-  sort_order: number;
-}
-
-export interface StashVersion {
-  id: string;
-  stash_id: string;
-  name: string;
-  description: string;
-  tags: string[];
-  metadata: Record<string, unknown>;
-  version: number;
-  created_by: string;
-  created_at: string;
-  files: StashVersionFile[];
-}
-
-export interface StashVersionListItem {
-  id: string;
-  stash_id: string;
-  name: string;
-  description: string;
-  version: number;
-  created_by: string;
-  created_at: string;
-  file_count: number;
-  total_size: number;
-}
-
-export interface StashFileInfo {
-  filename: string;
-  language: string;
-  size: number;
-}
-
-export interface StashListItem {
-  id: string;
-  name: string;
-  description: string;
-  tags: string[];
-  version: number;
-  archived: boolean;
-  created_at: string;
-  updated_at: string;
-  total_size: number;
-  files: StashFileInfo[];
-}
-
-export interface StashMeta {
-  id: string;
-  name: string;
-  description: string;
-  tags: string[];
-  metadata: Record<string, unknown>;
-  version: number;
-  archived: boolean;
-  created_at: string;
-  updated_at: string;
-  total_size: number;
-  files: StashFileInfo[];
-}
-
-export interface AccessLogEntry {
-  id: string;
-  stash_id: string;
-  source: 'api' | 'mcp' | 'ui';
-  action: string;
-  timestamp: string;
-  ip?: string;
-  user_agent?: string;
-}
-
-export type TokenScope = 'read' | 'write' | 'admin' | 'mcp';
-
-export interface ApiToken {
-  id: string;
-  label: string;
-  token_hash: string;
-  token_prefix: string;
-  scopes: TokenScope[];
-  created_at: string;
-}
-
-export interface ApiTokenListItem {
-  id: string;
-  label: string;
-  tokenPrefix: string;
-  scopes: TokenScope[];
-  createdAt: string;
-}
-
-export interface CreateStashInput {
-  name?: string;
-  description?: string;
-  tags?: string[];
-  metadata?: Record<string, unknown>;
-  files: { filename: string; content: string; language?: string }[];
-}
-
-export interface UpdateStashInput {
-  name?: string;
-  description?: string;
-  tags?: string[];
-  metadata?: Record<string, unknown>;
-  files?: { filename: string; content: string; language?: string }[];
-  // When set alongside content fields, the archive flag flip is applied
-  // inside the SAME transaction as the content update, so a thrown content
-  // update cannot leave the archive flag flipped (or vice-versa). When set
-  // alone, callers should use `archiveStash()` instead — that path skips
-  // the version snapshot per the documented archive semantics.
-  archived?: boolean;
-}
-
-export interface ListStashesOptions {
-  search?: string;
-  tag?: string;
-  archived?: boolean;
-  page?: number;
-  limit?: number;
-}
-
-export interface SearchStashItem extends StashListItem {
-  relevance: number;
-  snippets?: Record<string, string>;
-}
-
-export interface SearchStashesResult {
-  stashes: SearchStashItem[];
-  total: number;
-  query: string;
-}
-
-export interface TagGraphOptions {
-  tag?: string;
-  depth?: number;
-  min_weight?: number;
-  min_count?: number;
-  limit?: number;
-}
-
-export interface TagGraphResult {
-  nodes: { tag: string; count: number }[];
-  edges: { source: string; target: string; weight: number }[];
-  stash_count: number;
-  filter?: { tag: string; depth: number };
-}
-
-export interface StashGraphOptions {
-  mode?: 'relations' | 'timeline' | 'versions';
-  since?: string;
-  until?: string;
-  tag?: string;
-  limit?: number;
-  include_versions?: boolean;
-  min_shared_tags?: number;
-}
-
-export interface StashGraphNode {
-  id: string;
-  type: 'stash' | 'tag' | 'version';
-  label: string;
-  created_at?: string;
-  updated_at?: string;
-  version?: number;
-  file_count?: number;
-  total_size?: number;
-  tags?: string[];
-  count?: number;
-  version_number?: number;
-  created_by?: string;
-  change_summary?: Record<string, unknown>;
-}
-
-export interface StashGraphEdge {
-  source: string;
-  target: string;
-  type: 'has_tag' | 'shared_tags' | 'version_of' | 'temporal_proximity';
-  weight: number;
-  metadata?: {
-    shared_tags?: string[];
-    time_delta_hours?: number;
-  };
-}
-
-export interface StashGraphResult {
-  nodes: StashGraphNode[];
-  edges: StashGraphEdge[];
-  time_range: { min: string; max: string };
-  total_stashes: number;
-}
-
-interface Migration {
-  version: number;
-  name: string;
-  up: (db: Database.Database) => void;
-}
-
-const MIGRATIONS: Migration[] = [
-  {
-    version: 1,
-    name: 'add_stashes_version_column',
-    up: (db) => {
-      const columns = db.prepare("PRAGMA table_info(stashes)").all() as { name: string }[];
-      if (!columns.some(c => c.name === 'version')) {
-        db.exec('ALTER TABLE stashes ADD COLUMN version INTEGER NOT NULL DEFAULT 1');
-      }
-    }
-  },
-  {
-    version: 2,
-    name: 'add_stash_change_summary',
-    up: (db) => {
-      const columns = db.prepare("PRAGMA table_info(stash_versions)").all() as { name: string }[];
-      if (!columns.some(c => c.name === 'change_summary')) {
-        db.exec("ALTER TABLE stash_versions ADD COLUMN change_summary TEXT NOT NULL DEFAULT '{}'");
-      }
-    }
-  },
-  {
-    version: 3,
-    name: 'add_stash_graph_indexes',
-    up: (db) => {
-      db.exec(`
-        CREATE INDEX IF NOT EXISTS idx_stashes_tags ON stashes(tags);
-        CREATE INDEX IF NOT EXISTS idx_stash_versions_created_at ON stash_versions(created_at);
-      `);
-    }
-  },
-  {
-    version: 4,
-    name: 'add_stash_relations_table',
-    up: (db) => {
-      db.exec(`
-        CREATE TABLE IF NOT EXISTS stash_relations (
-          id TEXT PRIMARY KEY,
-          source_stash_id TEXT NOT NULL REFERENCES stashes(id) ON DELETE CASCADE,
-          target_stash_id TEXT NOT NULL REFERENCES stashes(id) ON DELETE CASCADE,
-          relation_type TEXT NOT NULL CHECK(relation_type IN ('shared_tags', 'temporal', 'manual')),
-          weight REAL NOT NULL DEFAULT 1.0,
-          metadata TEXT NOT NULL DEFAULT '{}',
-          created_at TEXT NOT NULL DEFAULT (datetime('now')),
-          UNIQUE(source_stash_id, target_stash_id, relation_type)
-        );
-
-        CREATE INDEX IF NOT EXISTS idx_stash_relations_source ON stash_relations(source_stash_id);
-        CREATE INDEX IF NOT EXISTS idx_stash_relations_target ON stash_relations(target_stash_id);
-        CREATE INDEX IF NOT EXISTS idx_stash_relations_type ON stash_relations(relation_type);
-      `);
-    }
-  },
-  {
-    version: 5,
-    name: 'add_graph_cache_table',
-    up: (db) => {
-      db.exec(`
-        CREATE TABLE IF NOT EXISTS graph_cache (
-          cache_key TEXT PRIMARY KEY,
-          data TEXT NOT NULL,
-          created_at TEXT NOT NULL DEFAULT (datetime('now')),
-          expires_at TEXT NOT NULL
-        );
-
-        CREATE INDEX IF NOT EXISTS idx_graph_cache_expires ON graph_cache(expires_at);
-      `);
-    }
-  },
-  {
-    version: 6,
-    name: 'backfill_stash_relations',
-    up: (db) => {
-      const stashes = db.prepare('SELECT id, tags FROM stashes').all() as { id: string; tags: string }[];
-      const parsed: { id: string; tags: string[] }[] = [];
-      for (const s of stashes) {
-        try { parsed.push({ id: s.id, tags: JSON.parse(s.tags) as string[] }); }
-        catch { /* skip stashes with corrupted tags JSON */ }
-      }
-
-      const insert = db.prepare(`
-        INSERT OR IGNORE INTO stash_relations (id, source_stash_id, target_stash_id, relation_type, weight, metadata)
-        VALUES (?, ?, ?, 'shared_tags', ?, ?)
-      `);
-
-      for (let i = 0; i < parsed.length; i++) {
-        for (let j = i + 1; j < parsed.length; j++) {
-          const shared = parsed[i].tags.filter(t => parsed[j].tags.includes(t));
-          if (shared.length > 0) {
-            const [src, tgt] = [parsed[i].id, parsed[j].id].sort();
-            insert.run(uuidv4(), src, tgt, shared.length, JSON.stringify({ shared_tags: shared }));
-          }
-        }
-      }
-    }
-  },
-  {
-    version: 7,
-    name: 'backfill_initial_version_records',
-    up: (db) => {
-      // Find stashes that have NO version records at all (never updated, or created before versioning)
-      const stashes = db.prepare(`
-        SELECT s.id, s.name, s.description, s.tags, s.metadata, s.created_at
-        FROM stashes s
-        WHERE NOT EXISTS (
-          SELECT 1 FROM stash_versions sv WHERE sv.stash_id = s.id
-        )
-      `).all() as { id: string; name: string; description: string; tags: string; metadata: string; created_at: string }[];
-
-      if (stashes.length === 0) return;
-
-      const insertVersion = db.prepare(`
-        INSERT INTO stash_versions (id, stash_id, name, description, tags, metadata, version, created_by, created_at, change_summary)
-        VALUES (?, ?, ?, ?, ?, ?, 1, 'system', ?, '{}')
-      `);
-
-      const insertVersionFile = db.prepare(`
-        INSERT INTO stash_version_files (id, version_id, filename, content, language, sort_order)
-        VALUES (?, ?, ?, ?, ?, ?)
-      `);
-
-      for (const s of stashes) {
-        const versionId = uuidv4();
-        insertVersion.run(versionId, s.id, s.name, s.description, s.tags, s.metadata, s.created_at);
-
-        const files = db.prepare('SELECT filename, content, language, sort_order FROM stash_files WHERE stash_id = ? ORDER BY sort_order').all(s.id) as { filename: string; content: string; language: string; sort_order: number }[];
-        for (const f of files) {
-          insertVersionFile.run(uuidv4(), versionId, f.filename, f.content, f.language, f.sort_order);
-        }
-      }
-
-      console.log(`[DB] Backfilled initial version records for ${stashes.length} stash(es)`);
-    }
-  },
-  {
-    version: 8,
-    name: 'add_fts5_search_index',
-    up: (db) => {
-      // Create FTS5 virtual table for ranked full-text search
-      db.exec(`
-        CREATE VIRTUAL TABLE IF NOT EXISTS stashes_fts USING fts5(
-          stash_id UNINDEXED,
-          name,
-          description,
-          tags,
-          filenames,
-          file_content,
-          tokenize='porter unicode61'
-        );
-      `);
-
-      // Backfill existing stashes
-      const stashes = db.prepare('SELECT id, name, description, tags FROM stashes').all() as {
-        id: string; name: string; description: string; tags: string;
-      }[];
-
-      const insertFts = db.prepare(`
-        INSERT INTO stashes_fts (stash_id, name, description, tags, filenames, file_content)
-        VALUES (?, ?, ?, ?, ?, ?)
-      `);
-
-      for (const s of stashes) {
-        const files = db.prepare('SELECT filename, content FROM stash_files WHERE stash_id = ? ORDER BY sort_order').all(s.id) as {
-          filename: string; content: string;
-        }[];
-        const filenames = files.map(f => f.filename).join(' ');
-        const fileContent = files.map(f => f.content).join('\n');
-        // Defensive parse — a corrupt tags row should not abort the migration
-        let tags = '';
-        try {
-          const parsed = JSON.parse(s.tags);
-          if (Array.isArray(parsed)) tags = parsed.filter((t: unknown) => typeof t === 'string').join(' ');
-        } catch { /* leave tags empty */ }
-        insertFts.run(s.id, s.name, s.description, tags, filenames, fileContent);
-      }
-
-      console.log(`[DB] FTS5 search index created and backfilled for ${stashes.length} stash(es)`);
-    }
-  },
-  {
-    version: 9,
-    name: 'add_stashes_archived_column',
-    up: (db) => {
-      const columns = db.prepare("PRAGMA table_info(stashes)").all() as { name: string }[];
-      if (!columns.some(c => c.name === 'archived')) {
-        db.exec('ALTER TABLE stashes ADD COLUMN archived INTEGER NOT NULL DEFAULT 0');
-        db.exec('CREATE INDEX IF NOT EXISTS idx_stashes_archived ON stashes(archived)');
-      }
-    }
-  },
-];
+// Re-export the data model so existing importers (`import { Stash } from
+// './db'`) keep working unchanged. The actual declarations now live in
+// db-types.ts. See refactor PR for #129.
+export type {
+  Stash,
+  StashFile,
+  StashVersionFile,
+  StashVersion,
+  StashVersionListItem,
+  StashFileInfo,
+  StashListItem,
+  StashMeta,
+  AccessLogEntry,
+  TokenScope,
+  ApiToken,
+  ApiTokenListItem,
+  CreateStashInput,
+  UpdateStashInput,
+  ListStashesOptions,
+  SearchStashItem,
+  SearchStashesResult,
+  TagGraphOptions,
+  TagGraphResult,
+  StashGraphOptions,
+  StashGraphNode,
+  StashGraphEdge,
+  StashGraphResult,
+} from './db-types';
 
 export class ClawStashDB {
   private db: Database.Database;
@@ -435,124 +77,8 @@ export class ClawStashDB {
   }
 
   private init() {
-    this.db.exec(`
-      CREATE TABLE IF NOT EXISTS stashes (
-        id TEXT PRIMARY KEY,
-        name TEXT NOT NULL DEFAULT '',
-        description TEXT NOT NULL DEFAULT '',
-        tags TEXT NOT NULL DEFAULT '[]',
-        metadata TEXT NOT NULL DEFAULT '{}',
-        created_at TEXT NOT NULL DEFAULT (datetime('now')),
-        updated_at TEXT NOT NULL DEFAULT (datetime('now'))
-      );
-
-      CREATE TABLE IF NOT EXISTS stash_files (
-        id TEXT PRIMARY KEY,
-        stash_id TEXT NOT NULL REFERENCES stashes(id) ON DELETE CASCADE,
-        filename TEXT NOT NULL,
-        content TEXT NOT NULL DEFAULT '',
-        language TEXT NOT NULL DEFAULT '',
-        sort_order INTEGER NOT NULL DEFAULT 0
-      );
-
-      CREATE TABLE IF NOT EXISTS access_log (
-        id TEXT PRIMARY KEY,
-        stash_id TEXT NOT NULL REFERENCES stashes(id) ON DELETE CASCADE,
-        source TEXT NOT NULL CHECK(source IN ('api', 'mcp', 'ui')),
-        action TEXT NOT NULL DEFAULT 'read',
-        timestamp TEXT NOT NULL DEFAULT (datetime('now')),
-        ip TEXT,
-        user_agent TEXT
-      );
-
-      CREATE TABLE IF NOT EXISTS api_tokens (
-        id TEXT PRIMARY KEY,
-        label TEXT NOT NULL DEFAULT '',
-        token_hash TEXT NOT NULL UNIQUE,
-        token_prefix TEXT NOT NULL,
-        scopes TEXT NOT NULL DEFAULT '["read"]',
-        created_at TEXT NOT NULL DEFAULT (datetime('now'))
-      );
-
-      CREATE INDEX IF NOT EXISTS idx_stashes_created_at ON stashes(created_at);
-      CREATE INDEX IF NOT EXISTS idx_stashes_updated_at ON stashes(updated_at);
-      CREATE INDEX IF NOT EXISTS idx_stash_files_stash_id ON stash_files(stash_id);
-      CREATE INDEX IF NOT EXISTS idx_access_log_stash_id ON access_log(stash_id);
-      CREATE INDEX IF NOT EXISTS idx_access_log_timestamp ON access_log(timestamp);
-      CREATE INDEX IF NOT EXISTS idx_api_tokens_hash ON api_tokens(token_hash);
-
-      CREATE TABLE IF NOT EXISTS admin_sessions (
-        id TEXT PRIMARY KEY,
-        token_hash TEXT NOT NULL UNIQUE,
-        created_at TEXT NOT NULL DEFAULT (datetime('now')),
-        expires_at TEXT
-      );
-
-      CREATE INDEX IF NOT EXISTS idx_admin_sessions_hash ON admin_sessions(token_hash);
-
-      CREATE TABLE IF NOT EXISTS stash_versions (
-        id TEXT PRIMARY KEY,
-        stash_id TEXT NOT NULL REFERENCES stashes(id) ON DELETE CASCADE,
-        name TEXT NOT NULL DEFAULT '',
-        description TEXT NOT NULL DEFAULT '',
-        tags TEXT NOT NULL DEFAULT '[]',
-        metadata TEXT NOT NULL DEFAULT '{}',
-        version INTEGER NOT NULL,
-        created_by TEXT NOT NULL DEFAULT '',
-        created_at TEXT NOT NULL DEFAULT (datetime('now'))
-      );
-
-      CREATE TABLE IF NOT EXISTS stash_version_files (
-        id TEXT PRIMARY KEY,
-        version_id TEXT NOT NULL REFERENCES stash_versions(id) ON DELETE CASCADE,
-        filename TEXT NOT NULL,
-        content TEXT NOT NULL DEFAULT '',
-        language TEXT NOT NULL DEFAULT '',
-        sort_order INTEGER NOT NULL DEFAULT 0
-      );
-
-      CREATE INDEX IF NOT EXISTS idx_stash_versions_stash_id ON stash_versions(stash_id);
-      CREATE INDEX IF NOT EXISTS idx_stash_versions_version ON stash_versions(stash_id, version);
-      CREATE INDEX IF NOT EXISTS idx_stash_version_files_version_id ON stash_version_files(version_id);
-    `);
-
-    this.migrate();
-  }
-
-  private migrate() {
-    // Ensure schema_migrations table exists
-    this.db.exec(`
-      CREATE TABLE IF NOT EXISTS schema_migrations (
-        version INTEGER PRIMARY KEY,
-        name TEXT NOT NULL,
-        applied_at TEXT NOT NULL DEFAULT (datetime('now'))
-      );
-    `);
-
-    // Load already-applied versions
-    const applied = new Set(
-      (this.db.prepare('SELECT version FROM schema_migrations').all() as { version: number }[])
-        .map(r => r.version)
-    );
-
-    // Run pending migrations in a transaction
-    const pending = MIGRATIONS.filter(m => !applied.has(m.version))
-      .sort((a, b) => a.version - b.version);
-
-    if (pending.length === 0) return;
-
-    const runMigrations = this.db.transaction(() => {
-      for (const migration of pending) {
-        console.log(`[DB] Running migration ${migration.version}: ${migration.name}`);
-        migration.up(this.db);
-        this.db.prepare(
-          'INSERT INTO schema_migrations (version, name, applied_at) VALUES (?, ?, ?)'
-        ).run(migration.version, migration.name, new Date().toISOString());
-      }
-    });
-
-    runMigrations();
-    console.log(`[DB] ${pending.length} migration(s) applied successfully`);
+    this.db.exec(INITIAL_SCHEMA_SQL);
+    applyPendingMigrations(this.db);
   }
 
   private safeParseTags(raw: unknown): string[] {
